review comments; pass 1
